requirements.txt
cassandra-driver==3.25.0
pandas==1.3.3